additional weather detail fragments, fully functional upper and middle bars

+++ app/src/main/java/com/vvash17company/weatherapp/models/apiservicemodels/Wind.kt
@@ -3,5 +3,5 @@ package com.vvash17company.weatherapp.models.apiservicemodels
 import com.google.gson.annotations.SerializedName
 
 class Wind(
-    @SerializedName("speed") var speed: Long
+    @SerializedName("speed") var speed: Double
 )
